+++ src/components/Home/component/PrimaryActions/PointRechargeButton/PointRechargeButton.js
@@ -8,6 +8,7 @@ import {
   servicesHandler,
   SERVICES_TYPE
 } from '../../../../../helper/servicesHandler';
+import SubActionButton from '../SubActionButton';
 
 const styles = StyleSheet.create({
   add_store_action_btn: {
@@ -31,7 +32,7 @@ const PointRechargeButton = ({
   label,
   wrapperStyle,
   containerStyle,
-  iconName = 'plus-circle',
+  iconName = 'ios-add-circle',
   iconStyle,
   labelStyle,
   onPress = () => {}
@@ -49,19 +50,16 @@ const PointRechargeButton = ({
   }
 
   return (
-    <TouchableOpacity
-      onPress={goPaymentMethod}
-      style={[styles.add_store_action_btn, wrapperStyle]}
-    >
-      <View style={[styles.add_store_action_btn_box, containerStyle]}>
-        <Icon name={iconName} style={[styles.icon, iconStyle]} />
-        {!!label && (
-          <Text style={[styles.add_store_action_label, labelStyle]}>
-            {label}
-          </Text>
-        )}
-      </View>
-    </TouchableOpacity>
+    <SubActionButton 
+    iconName={iconName}
+    label={label}
+
+    wrapperStyle={wrapperStyle}
+    containerStyle={containerStyle}
+    iconStyle={iconStyle}
+
+    onPress={goPaymentMethod}
+    />
   );
 };